+++ R/export.R
@@ -1,10 +1,10 @@
 #' @rdname export
 #' @title Export
 #' @description Write data.frame to a file
-#' @param x A data frame or matrix to be written into a file.
+#' @param x A data frame or matrix to be written into a file. (An exception to this is that \code{x} can be a list of data frames if the output file format is an Excel .xlsx workbook. See examples.)
 #' @param file A character string naming a file. Must specify \code{file} and/or \code{format}.
 #' @param format An optional character string containing the file format, which can be used to override the format inferred from \code{file} or, in lieu of specifying \code{file}, a file with the symbol name of \code{x} and the specified file extension will be created. Must specify \code{file} and/or \code{format}. Shortcuts include: \dQuote{,} (for comma-separated values), \dQuote{;} (for semicolon-separated values), \dQuote{|} (for pipe-separated values), and \dQuote{dump} for \code{\link[base]{dump}}.
-#' @param \dots Additional arguments for the underlying export functions.
+#' @param \dots Additional arguments for the underlying export functions. See examples.
 #' @return The name of the output file as a character string (invisibly).
 #' @details This function exports a data frame or matrix into a file with file format based on the file extension (or the manually specified format, if \code{format} is specified).
 #'
@@ -31,7 +31,7 @@
 #'     \item Weka Attribute-Relation File Format (.arff), using \code{\link[foreign]{write.arff}}
 #'     \item R syntax object (.R), using \code{\link[base]{dput}} (by default) or \code{\link[base]{dump}} (if \code{format = 'dump'}
 #'     \item Matlab (.mat), using \code{\link[rmatio]{write.mat}}
-#'     \item Excel (.xlsx), using \code{\link[openxlsx]{write.xlsx}}
+#'     \item Excel (.xlsx), using \code{\link[openxlsx]{write.xlsx}}. Use \code{which} to specify a sheet name and \code{overwrite} to decide whether to overwrite an existing file or worksheet (the default) or add the data as a new worksheet (with \code{overwrite = FALSE}). \code{x} can also be a list of data frames; the list entry names are used as sheet names.
 #'     \item OpenDocument Spreadsheet (.ods), using \code{\link[readODS]{write_ods}}. (Currently only single-sheet exports are supported.)
 #'     \item XML (.xml), using a custom method based on \code{\link[xml2]{xml_add_child}} to create a simple XML tree and \code{\link[xml2]{write_xml}} to write to disk.
 #'     \item HTML (.html), using a custom method based on \code{\link[xml2]{xml_add_child}} to create a simple HTML table and \code{\link[xml2]{write_xml}} to write to disk.
@@ -65,6 +65,18 @@
 #' export(mtcars, file = "data.R", format = "dump")
 #' export(mtcars, file = "data.R", format = "dump", append = TRUE)
 #' source("data.R", echo = TRUE)
+#'
+#' # write to an Excel workbook
+#' \dontrun{
+#'   ## export a single data frame
+#'   export(mtcars, "mtcars.xlsx")
+#'   
+#'   ## export a list of data frames as worksheets
+#'   export(list(a = mtcars, b = iris), "multisheet.xlsx")
+#' 
+#'   ## export, adding sheet to an existing workbook
+#'   export(iris, "mtcars.xlsx", which = "iris", overwrite = FALSE)
+#' }
 #'
 #' # write data to a zip-compressed CSV
 #' export(mtcars, "mtcars.csv.zip")
@@ -103,8 +115,11 @@ export <- function(x, file, format, ...) {
         on.exit(close(file))
     }
 
+    data_name <- as.character(substitute(x))
     if (!is.data.frame(x) & !is.matrix(x)) {
-        stop("'x' is not a data.frame or matrix")
+        if (fmt != "xlsx") {
+            stop("'x' is not a data.frame or matrix")
+        }
     } else if (is.matrix(x)) {
         x <- as.data.frame(x)
     }